moved recipe list mapping from RecipeList to MainComponent
